+++ src/Middleware/ErrorCatcher.php
@@ -15,6 +15,7 @@ use Psr\Http\Server\RequestHandlerInterface;
 use Throwable;
 use Yiisoft\ErrorHandler\Event\ApplicationError;
 use Yiisoft\ErrorHandler\ErrorHandler;
+use Yiisoft\ErrorHandler\HeadersProvider;
 use Yiisoft\ErrorHandler\Renderer\HeaderRenderer;
 use Yiisoft\ErrorHandler\Renderer\HtmlRenderer;
 use Yiisoft\ErrorHandler\Renderer\JsonRenderer;
@@ -39,6 +40,8 @@ use function trim;
  */
 final class ErrorCatcher implements MiddlewareInterface
 {
+    private HeadersProvider $headersProvider;
+
     /**
      * @psalm-var array<string,class-string<ThrowableRendererInterface>>
      */
@@ -57,7 +60,9 @@ final class ErrorCatcher implements MiddlewareInterface
         private ErrorHandler $errorHandler,
         private ContainerInterface $container,
         private ?EventDispatcherInterface $eventDispatcher = null,
+        HeadersProvider $headersProvider = null,
     ) {
+        $this->headersProvider = $headersProvider ?? new HeadersProvider();
     }
 
     /**
@@ -143,7 +148,9 @@ final class ErrorCatcher implements MiddlewareInterface
 
         $data = $this->errorHandler->handle($t, $renderer, $request);
         $response = $this->responseFactory->createResponse(Status::INTERNAL_SERVER_ERROR);
-
+        foreach ($this->headersProvider->getAll() as $name => $value) {
+            $response = $response->withHeader($name, $value);
+        }
         return $data->addToResponse($response->withHeader(Header::CONTENT_TYPE, $contentType));
     }